# openapi.json
{
  "components": {
    "schemas": {
...
      "DismissRequest": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
...
      "MergeCandidateView": {
        "description": "A pending merge candidate enriched with both series' display titles, for the operator\nreview queue (design §11 `GET /v1/admin/merge-candidates`).",
        "properties": {
          "candidate_id": {
            "$ref": "#/components/schemas/SeriesId"
          },
          "candidate_title": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "score": {
            "format": "float",
            "type": "number"
          },
          "series_id": {
            "$ref": "#/components/schemas/SeriesId"
          },
          "series_title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "series_id",
          "series_title",
          "candidate_id",
          "candidate_title",
          "score",
          "created_at"
        ],
        "type": "object"
      },
      "MergeRequest": {
        "properties": {
          "keep": {
            "$ref": "#/components/schemas/SeriesId",
            "description": "The surviving canonical series."
          },
          "merge": {
            "$ref": "#/components/schemas/SeriesId",
            "description": "The series merged into `keep` and then deleted."
          }
        },
        "required": [
          "keep",
          "merge"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "RFC 9457 `application/problem+json` error body shape produced by [`ApiError`]. Declared\npurely for `OpenAPI` documentation — [`ApiError::into_response`] builds the JSON by hand so\nruntime callers never construct this type.",
        "example": {
          "detail": "resource not found",
          "status": 404,
          "title": "not_found",
          "type": "about:blank#not_found"
        },
        "properties": {
          "detail": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "type": "object"
      },
...
      "SeriesId": {
        "description": "Identifies a canonical [`crate::entities::Series`].",
        "format": "uuid",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "TankoVault contributors"
    },
    "description": "Axum HTTP edge: public read/write + admin, auth, resolved links, SSE.",
    "license": {
      "name": ""
    },
    "title": "tankovault-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/merge-candidates": {
      "get": {
        "description": "The canonicalisation review queue (design §10).",
        "operationId": "list_merge_candidates",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MergeCandidateView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Up to 200 open merge candidates"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "caller must have at least the operator role"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "List merge candidates",
        "tags": [
          "admin-matching"
        ]
      }
    },
    "/v1/admin/merge-candidates/dismiss": {
      "post": {
        "description": "Operator judged the two works distinct.",
        "operationId": "dismiss_merge_candidate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DismissRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "dismissed": true
                },
                "schema": {}
              }
            },
            "description": "Whether a candidate was actually dismissed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "caller must have at least the operator role"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Dismiss a merge candidate",
        "tags": [
          "admin-matching"
        ]
      }
    },
...
    "/v1/admin/series/merge": {
      "post": {
        "description": "Transactional re-parent + title/tag union (design §10).",
        "operationId": "merge_series",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "ok": true
                },
                "schema": {}
              }
            },
            "description": "Acknowledged"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "caller must have at least the operator role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "One or both series not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "`keep` and `merge` are the same series"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Merge two series",
        "tags": [
          "admin-matching"
        ]
      }
    },
...
  },
...
}